Clean up the merging POC and add light docs
