+++ compiler-k2-utils/src/main/kotlin/com/squareup/anvil/compiler/k2/utils/CollectionExtensions.kt
@@ -1,6 +1,4 @@
-package com.squareup.anvil.compiler.k2.fir.internal
-
-import com.squareup.anvil.compiler.k2.fir.singleOrEmpty
+package com.squareup.anvil.compiler.k2.utils
 
 private val truePredicate: (Any?) -> Boolean = { true }
 
@@ -8,7 +6,28 @@ private val truePredicate: (Any?) -> Boolean = { true }
  * Returns single element, or `null` if the collection is empty. Unlike [singleOrNull] this
  * method throws an exception if more than one element is found.
  */
-internal fun <T> Iterable<T>.singleOrEmpty(): T? = singleOrEmpty(truePredicate)
+public fun <T> Iterable<T>.singleOrNone(): T? = singleOrNone(truePredicate)
+
+/**
+ * Returns the single element matching the given [predicate], or `null` if element was not found.
+ * Unlike [singleOrNull] this method throws an exception if more than one element is found.
+ */
+public inline fun <T> Iterable<T>.singleOrNone(predicate: (T) -> Boolean): T? {
+  var single: T? = null
+  var found = false
+  for (element in this) {
+    if (predicate(element)) {
+      if (found) {
+        throw IllegalArgumentException(
+          "Collection contains more than one matching element.",
+        )
+      }
+      single = element
+      found = true
+    }
+  }
+  return single
+}
 
 /**
  * Transforms the elements of the receiver collection and adds the results to a set.
@@ -19,7 +38,7 @@ internal fun <T> Iterable<T>.singleOrEmpty(): T? = singleOrEmpty(truePredicate)
  * @receiver The collection to be transformed.
  * @return A set containing the transformed elements from the receiver collection.
  */
-internal inline fun <C : Collection<T>, T, R> C.mapToSet(
+public inline fun <C : Collection<T>, T, R> C.mapToSet(
   destination: MutableSet<R> = mutableSetOf(),
   transform: (T) -> R,
 ): Set<R> = mapTo(destination, transform)
@@ -47,7 +66,7 @@ public inline fun <C : Collection<T>, T, R : Any> C.mapNotNullToSet(
  * @receiver The array to be transformed.
  * @return A set containing the transformed elements from the receiver array.
  */
-internal inline fun <T, R> Array<T>.mapToSet(
+public inline fun <T, R> Array<T>.mapToSet(
   destination: MutableSet<R> = mutableSetOf(),
   transform: (T) -> R,
 ): Set<R> {
