+++ compiler-k2/build.gradle.kts
@@ -7,9 +7,6 @@ plugins {
 conventions {
   explicitApi = true
   kotlinCompilerArgs.addAll(
-    // The flag is needed because we extend an interface that uses @JvmDefault and the Kotlin
-    // compiler requires this flag when doing so.
-    "-Xjvm-default=all",
     "-opt-in=org.jetbrains.kotlin.compiler.plugin.ExperimentalCompilerApi",
   )
 }
@@ -29,8 +26,6 @@ dependencies {
   api(project(":compiler-k2-api"))
   api(project(":compiler-k2-utils"))
   api(libs.kotlin.compiler)
-  api(libs.inject)
-  api(libs.dagger2)
 
   kapt(libs.auto.service.processor)
